Update BMI_GUI
The GUI-Based BMI Calculator is a Python application using Tkinter to calculate and categorize BMI based on user input of height and weight, saving these details to a CSV file. The user-friendly interface instantly displays the BMI and corresponding category and uses Matplotlib to display the pie chart of BMI categories.